some folder restructuring
